+++ src/action/action-orchestrator.ts
@@ -82,7 +82,7 @@ export class ActionOrchestrator {
         reportResult = await reportGenerator.generateReport(reportData, {
           maxSize: reporter.maxSize ?? undefined,
           // eslint-disable-next-line @typescript-eslint/no-extra-non-null-assertion,@typescript-eslint/no-non-null-assertion
-          failOnSeverities: this.inputs!!.failOnSeverities
+          considerErrorOnSeverities: this.inputs!!.considerErrorOnSeverities
         })
         reportResults.set(reporter.maxSize, reportResult)
       }
@@ -103,7 +103,7 @@ export class ActionOrchestrator {
 
       const failed = await this.doReports(report, reporters)
 
-      return failed ? 1 : 0
+      return failed && this.inputs.failOnError ? 1 : 0
     } catch (e) {
       this.gitHubCheck?.cancel()
       throw e