+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2023-09-19
+
+### Changed
+
+- Remove `fail-on-severities` and separate in two flags
+  - Add `consider-error-on-severities` to consider specific severities as errors
+  - Re-add `fail-on-error` to fail the action if an error occurs
+
 ## [2.0.0] - 2023-09-19
 
 ### Added
@@ -62,7 +70,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement `sechub-reporter-action` github action
 
-[Unreleased]: https://github.com/tvcsantos/sechub-reporter-action/compare/v2.0.0...main
+[Unreleased]: https://github.com/tvcsantos/sechub-reporter-action/compare/v3.0.0...main
+[3.0.0]: https://github.com/tvcsantos/sechub-reporter-action/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/tvcsantos/sechub-reporter-action/compare/v1.4.1...v2.0.0
 [1.4.1]: https://github.com/tvcsantos/sechub-reporter-action/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/tvcsantos/sechub-reporter-action/compare/v1.3.0...v1.4.0